lab_pneumonia/streamlit/requirements_updated.txt
# Updated requirements with compatible versions for Python 3.11+
# Use this if the original requirements.txt fails

numpy>=1.24.0
streamlit>=1.22.0
Pillow>=9.5.0
keras>=2.12.0
tensorflow>=2.12.0

